requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
et-xmlfile==1.1.0
idna==3.10
openpyxl==3.1.5
peewee==3.17.7
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
